Having trouble generalizing
The problem is that fmap returns a wrapped value, and as such the
applicatives end up layered. To effectively do this a way to unwrap the
values is required, however that would penetrate the veil too much I
think.

Is it possible to lift the value out? AFAIK functor / applicative
provide no guarantee that it is (makes sense, as the Nothing has no
value to lift, and the array can have many - so what would it return).